Implementing //TODOs in PhoneBookImpl
Added utility methods to DatabauseUtil.java for printing database
contents and truncating the table.  These are used in unit tests.

addPerson() only adds the Person object to the local store
addPhoneBookContentsToDatabase() commits all entries from the local
store to the database.
findPerson() searches the local store and the database for the specified
Name.